[refactor] (ABC-1077): Move functions from scheduler to shared utils
