+++ src/modules/base/schedulerUtils/schedulerUtils.js
@@ -1,5 +1,3 @@
-
-
 export {
     containsAllowedDateTimes,
     getDisabledWeekdaysLabels,
@@ -10,7 +8,6 @@ export {
     nextAllowedDay,
     nextAllowedMonth,
     nextAllowedTime,
-    parseTimeFrame,
     previousAllowedDay,
     previousAllowedMonth,
     previousAllowedTime,
